项目基本完成

+++ warehouse-mgr-be/src/routers/goods/index.js
@@ -8,7 +8,7 @@ const GOODS_CONST = {
 };
 
 const Goods = mongoose.model('Goods');
-const InventoryLog = mongoose.model('InventoryLog');
+// const InventoryLog = mongoose.model('InventoryLog');
 
 const findGoodsOne = async (id) => {
     const one = await Goods.findOne({
@@ -24,15 +24,31 @@ const router = new Router({
 
 router.post('/add', async (ctx) => {
     const {
+        newId,
         name,
         price,
         supplier,
         launchDate,
         classify,
-        count,
     } = getBody(ctx);
 
+    let count = 0;
+
+    const newIdOne = await Goods.findOne({
+        newId: newId,
+    }).exec();
+
+    if (newIdOne) {
+        ctx.body = {
+            code: 0,
+            msg: '商品编号已存在',
+        };
+
+        return;
+    };
+
     const goods = new Goods({
+        newId,
         name,
         price,
         supplier,
@@ -57,7 +73,7 @@ router.get('/list', async (ctx) => {
         keyword = ''
     } = ctx.query;
 
-    let = {
+    let {
         size = 10,
     } = ctx.query;
 
@@ -72,7 +88,7 @@ router.get('/list', async (ctx) => {
     const list = await Goods
         .find(query)
         .sort({
-            _id: -1,
+            newId: -1,
         })
         .skip((page - 1) * size)
         .limit(size)
@@ -97,6 +113,19 @@ router.delete('/:id', async (ctx) => {
         id
     } = ctx.params;
 
+    const delJudge = await Goods.findOne({
+        _id: id,
+    }).exec();
+
+    if (delJudge.count > 0) {
+        ctx.body = {
+            code: 0,
+            msg: '该商品库存不为零，无法删除',
+        };
+
+        return;
+    };
+
     const delMsg = await Goods.deleteOne({
         _id: id,
     });
@@ -108,71 +137,77 @@ router.delete('/:id', async (ctx) => {
     };
 });
 
-router.post('/update/count', async (ctx) => {
-    const {
-        id,
-        type,
-    } = ctx.request.body;
+// router.post('/update/count', async (ctx) => {
+//     const {
+//         id,
+//         type,
+//     } = ctx.request.body;
 
-    let {
-        num,
-    } = ctx.request.body;
+//     let {
+//         num,
+//     } = ctx.request.body;
 
-    num = Number(num);
+//     num = Number(num);
 
-    const goods = await findGoodsOne(id);
+//     const goods = await findGoodsOne(id);
 
-    // 没有找到商品
-    if (!goods) {
-        ctx.body = {
-            code: 0,
-            msg: '没有找到商品',
-        };
+//     // 没有找到商品
+//     if (!goods) {
+//         ctx.body = {
+//             code: 0,
+//             msg: '没有找到商品',
+//         };
 
-        return;
-    };
+//         return;
+//     };
 
-    // 找到了商品
-    if (type === GOODS_CONST.IN) {
-        // 入库操作
-        num = Math.abs(num);
-    } else {
-        // 出库操作
-        num = - Math.abs(num);
-    };
+//     // 找到了商品
+//     if (type === GOODS_CONST.IN) {
+//         // 入库操作
+//         num = Math.abs(num);
+//     } else {
+//         // 出库操作
+//         num = - Math.abs(num);
+//     };
 
-    goods.count = goods.count + num;
+//     goods.count = goods.count + num;
 
-    if (goods.count < 0) {
-        ctx.body = {
-            code: 0,
-            msg: '剩下的量不足以出库',
-        }
+//     if (goods.count < 0) {
+//         ctx.body = {
+//             code: 0,
+//             msg: '剩下的量不足以出库',
+//         }
 
-        return;
-    };
+//         return;
+//     };
 
-    const res = await goods.save();
+//     const res = await goods.save();
 
-    const log = new InventoryLog({
-        num: Math.abs(num),
-        type,
-        goodsId: id,
-    });
+//     const log = new InventoryLog({
+//         num: Math.abs(num),
+//         type,
+//         goodsId: id,
+//     });
 
-    log.save();
+//     log.save();
 
-    ctx.body = {
-        data: res,
-        code: 1,
-        msg: '操作成功',
-    }
-});
+//     ctx.body = {
+//         data: res,
+//         code: 1,
+//         msg: '操作成功',
+//     }
+// });
 
 router.post('/update', async (ctx) => {
     const {
         id,
-        ...others
+        name,
+        price,
+        supplier,
+        launchDate,
+        classify,
+        specification,
+        unit,
     } = ctx.request.body;
 
     const one = await findGoodsOne(id);
@@ -189,10 +224,18 @@ router.post('/update', async (ctx) => {
     // 找到了商品
     const newQuery = {};
 
-    Object.entries(others).forEach(([key, value]) => {
+    Object.entries({
+        name,
+        price,
+        supplier,
+        launchDate,
+        classify,
+        specification,
+        unit,
+    }).forEach(([key, value]) => {
         if (value) {
             newQuery[key] = value;
-        }
+        };
     });
 
     Object.assign(one, newQuery);
@@ -203,7 +246,7 @@ router.post('/update', async (ctx) => {
         data: res,
         code: 1,
         msg: '保存成功',
-    }
+    };
 });
 
 router.get('/detail/:id', async (ctx) => {
@@ -211,7 +254,9 @@ router.get('/detail/:id', async (ctx) => {
         id,
     } = ctx.params;
 
-    const one = await findGoodsOne(id);
+    const one = await Goods.findOne({
+        newId: id,
+    });
 
     // 没有找到商品
     if (!one) {
@@ -226,6 +271,18 @@ router.get('/detail/:id', async (ctx) => {
         msg: '查询成功',
         data: one,
         code: 1,
+    };
+});
+
+router.get('/list/all', async (ctx) => {
+    const list = await Goods.find().sort({
+        _id: -1,
+    }).exec();
+
+    ctx.body = {
+        data: list,
+        code: 1,
+        msg: '获取成功',
     };
 });
